refactor: place network module into 'core' subfolder
