+++ src/core/socket.cpp
@@ -8,7 +8,7 @@
  * This is an implementation file implementing the Socket, which acts as either
  * the listener or client
  */
-#include "socket.h"
+#include "core/socket.h"
 
 #include <fcntl.h>
 #include <sys/socket.h>
@@ -18,7 +18,7 @@
 #include <cerrno>
 #include <stdexcept>
 
-#include "net_address.h"
+#include "core/net_address.h"
 namespace TURTLE_SERVER {
 
 static constexpr int BACK_LOG = 128;
