+++ src/include/core/connection.h
@@ -18,9 +18,9 @@
 #include <utility>
 #include <vector>
 
-#include "buffer.h"
-#include "socket.h"
-#include "utils.h"
+#include "core/buffer.h"
+#include "core/socket.h"
+#include "core/utils.h"
 
 #define TEMP_BUF_SIZE 2048
