+++ src/include/core/looper.h
@@ -19,7 +19,7 @@
 #include <memory>
 #include <mutex>  // NOLINT
 
-#include "utils.h"
+#include "core/utils.h"
 
 namespace TURTLE_SERVER {
