+++ src/include/core/thread_pool.h
@@ -21,7 +21,7 @@
 #include <utility>
 #include <vector>
 
-#include "utils.h"
+#include "core/utils.h"
 #ifndef SRC_INCLUDE_THREAD_POOL_H_
 #define SRC_INCLUDE_THREAD_POOL_H_
