+++ src/include/core/utils.h
